Time: 11 ms (41.24%) | Memory: 55.2 MB (69.21%) - LeetSync

+++ 1956-maximum-element-after-decreasing-and-rearranging/maximum-element-after-decreasing-and-rearranging.cpp
@@ -4,7 +4,7 @@ public:
     int maximumElementAfterDecrementingAndRearranging(vector<int>& arr) {
         sort(arr.begin(),arr.end());
         arr[0]=1;
-        for (int i = 1; i < arr.size(); i++)
+        for (int i = 1; i < (int)arr.size(); i++)
         {
             if (arr[i]>1+arr[i-1])
             {
